First step done for packaging
the main.py file in the root can run and import the code fine. Added some statements to __init__.py in pyEngine2D directory. Also changed the importing of data and scripts so now the builtins are wrapped in the package and and user ones can just be put anywhere near the main.py file.